修改bug

+++ 面试题44：数字序列中某一位的数字/数字序列中某一位的数字.cpp
@@ -14,7 +14,7 @@ public:
         if(n < 0)return -1;
         if(n < 10) return n;
         int base = 10, bitnums = 2;
-        int totalNums = 10, preTotalNums;
+        long long totalNums = 10, preTotalNums;
         int ans;
         while(true){
             preTotalNums = totalNums;
